+++ template/root/template/base/lib/index.cjs
@@ -1,3 +1,3 @@
-const sum = require("../base/dist/sum.cjs.min.js");
+const sum = require('../dist/sum.cjs.min.js');
 
 exports.sum = sum;