+++ template/test/index.js
@@ -1,7 +1,7 @@
 import path from "node:path";
 import { copyTmpl, copyFile } from "../../utils/copy.cjs";
-import { mergeTmpl2JSON } from "../../utils/merge.js";
-import { currentPathResolve } from "../../utils/helper.js";
+import { mergeTmpl2JSON, mergeJSON2JSON } from "../../utils/merge.js";
+import { currentPathResolve, checkFileExists } from "../../utils/helper.js";
 
 // 拼接当前目录下template的绝对路径
 function pathDirResolve(p) {
@@ -12,7 +12,7 @@ function pathDirResolve(p) {
 
 function initTest(cmdPath, name, option) {
   console.log(cmdPath, name, option);
-  if (!option.husky) {
+  if (!option.test) {
     return;
   }
 
@@ -25,6 +25,19 @@ function initTest(cmdPath, name, option) {
     );
     // 复制单元测试demo文件到tests根目录文件夹下
     copyFile(pathDirResolve(`tests/sum.spec.js`), path.resolve(cmdPath, name, "tests/sum.spec.js"));
+    if (checkFileExists(path.join(cmdPath + name, "/babel.config.json"))) {
+      // 如果存在就合并babel.config.json
+      mergeTmpl2JSON(
+        pathDirResolve(`babel.config.json.tmpl`),
+        path.resolve(cmdPath, name, "babel.config.json"),
+        option,
+      );
+    } else {
+      copyFile(
+        pathDirResolve(`babel.config.json`),
+        path.resolve(cmdPath, name, "babel.config.json"),
+      );
+    }
   } else if (option.test.Vitest) {
     // 选择的是vitest
     copyTmpl(